just make vagrantfile instead of calling init
